refactor: move createPerson to new service
